requirements.txt
beautifulsoup4==4.11.2
bokeh==2.4.3
geopandas==0.10.2
iqplot==0.3.2
matplotlib==3.5.1
numpy>=1.23.0
pandas==1.4.2
Pillow==9.4.0
pydeck==0.8.0b4
requests==2.28.1
seaborn==0.11.2
st_annotated_text==3.0.0
st_pages==0.3.1
streamlit==1.17.0
streamlit_extras==0.2.6
streamlit_lottie==0.0.3
xyzservices==2022.9.0
